requirements.txt
azure-storage-blob 
azure-identity
tqdm
python-dotenv==0.21.1